web/requirements.txt
asgiref==3.2.7
Django==3.0
django-cors-headers==3.2.1
django-extensions==2.2.9
djangorestframework==3.11.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
